refactor(api): extract output unit enum into common package
