+++ api/src/common/output/index.ts
@@ -1,0 +1,7 @@
+enum OutputUnit {
+    SECONDS = "SECONDS",
+    MINUTES = "MINUTES",
+    GAME_DAYS = "GAME_DAYS",
+}
+
+export { OutputUnit };
