+++ api/src/functions/query-output/domain/output-calculator.ts
@@ -3,13 +3,11 @@ import {
     getMaxToolModifier,
     isAvailableToolSufficient,
 } from "../../../common/modifiers";
+import { OutputUnit } from "../../../common/output";
 import { Tools } from "../../../types";
 import { queryOutputDetails } from "../adapters/mongodb-output-adapter";
 import { ItemOutputDetails } from "../interfaces/output-database-port";
-import type {
-    OutputUnit,
-    QueryOutputPrimaryPort,
-} from "../interfaces/query-output-primary-port";
+import type { QueryOutputPrimaryPort } from "../interfaces/query-output-primary-port";
 import OutputUnitSecondMappings from "../utils/OutputUnitSecondMapping";
 
 const INVALID_ITEM_NAME_ERROR =
