+++ api/src/functions/query-output/handler.ts
@@ -1,8 +1,8 @@
 import type { QueryOutputArgs } from "../../graphql/schema";
 import type { GraphQLEventHandler } from "../../interfaces/GraphQLEventHandler";
 import { calculateOutput } from "./domain/output-calculator";
-import { OutputUnit } from "./interfaces/query-output-primary-port";
 import { ToolSchemaMap } from "../../common/modifiers";
+import { OutputUnit } from "../../common/output";
 
 const handler: GraphQLEventHandler<QueryOutputArgs, number> = async (event) => {
     const { name, workers, unit, maxAvailableTool, creator } = event.arguments;
